Add new path files and update path planner settings
This commit adds new path files for the robot to follow and updates the path planner settings and relevant auto file. It also updates the version of PathplannerLib in the vendor dependencies to 2024.1.4. The paths "POS 1 To AMP", "E1 To Speaker Score", and others, have been added, while the file "Amp To Middle Note 1" has been modified.