Update PID constants and swerve drive configuration
The PID constant values for both drive and angle have been updated in the pidfproperties.json file to improve system response. In the SwerveSubsystem.java file, the creation of the swerve drive has been adjusted to use the new maximum speed. Also, the HolonomicPathFollowerConfig has been updated by setting the PID constants directly from the swerve drive configuration to keep consistency and clarity in the code.