Update autonomous command and tweak PID constants
Several adjustments were made to improve the robot's functionality. The autonomous command was updated to use a new path planner and to add a simple and complex auto. Additionally, tweaks were made to PID constants and controller properties to refine the robot's movements. The YAGSL version was updated for better swerve library compatibility.